+++ pyi.py
@@ -329,6 +329,12 @@ def _is_object_or_Unused(node: ast.expr | None) -> bool:
     return _is_builtins_object(node) or _is_Unused(node)
 
 
+def _is_IncompleteOrNone(node: ast.expr | None) -> bool:
+    if not isinstance(node, ast.BinOp) or not isinstance(node.op, ast.BitOr):
+        return False
+    return _is_Incomplete(node.left) and _is_None(node.right)
+
+
 def _get_name_of_class_if_from_modules(
     classnode: ast.expr, *, modules: Container[str]
 ) -> str | None:
@@ -383,7 +389,7 @@ def _is_type_or_Type(node: ast.expr) -> bool:
     return cls_name in {"type", "Type"}
 
 
-def _is_None(node: ast.expr) -> bool:
+def _is_None(node: ast.expr | None) -> bool:
     return isinstance(node, ast.Constant) and node.value is None
 
 
@@ -2226,6 +2232,8 @@ class PyiVisitor(ast.NodeVisitor):
                 self.visit(default)
         if default is not None and not _is_valid_default_value_with_annotation(default):
             self.error(default, (Y014 if arg.annotation is None else Y011))
+        if _is_IncompleteOrNone(arg.annotation) and _is_None(default):
+            self.error(arg, Y067)
 
     def error(self, node: NodeWithLocation, message: str) -> None:
         self.errors.append(Error(node.lineno, node.col_offset, message, PyiTreeChecker))
@@ -2429,6 +2437,7 @@ Y066 = (
     "Y066 When using if/else with sys.version_info, "
     'put the code for new Python versions first, e.g. "{new_syntax}"'
 )
+Y067 = 'Y067 Use "=None" instead of "Incomplete | None = None"'
 Y090 = (
     'Y090 "{original}" means '
     '"a tuple of length 1, in which the sole element is of type {typ!r}". '